pyproject.toml
[tool.poetry]
name = "drivethru"
version = "0.1.0"
description = ""
authors = ["Elio Kazu Mostero", "Corey Ryan Mostero"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.25.1"
sounddevice = "^0.4.6"
numpy = "^1.26.4"
speechrecognition = "^3.10.4"
pre-commit = "^3.7.0"
pyaudio = "^0.2.14"
python-dotenv = "^1.0.1"
pydub = "^0.25.1"
pyyaml = "^6.0.1"
types-pyyaml = "^6.0.12.20240311"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
mypy = "^1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
